# docs/web/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Admin REST API шины данных (§7 ТЗ). Сессии в Redis +\nAPI-токены (Bearer db_*). SPA по embed.FS отдаётся\nфолбэком на index.html для всех путей не из /api/.",
        "title": "Nexus Web API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/audit": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "ApiTokenAuth": []
                    }
                ],
                "description": "Поддерживаемые actions: node.create/update/delete, user.create/update/delete/password_changed, token.create/revoke/delete, login.success/failure, ch_table.drop, settings.update и др.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "Журнал audit-log (admin only, §7.13).",
                "parameters": [
                    {
                        "type": "string",
                        "description": "фильтр по user_id",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "одно значение action",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "несколько action через запятую",
                        "name": "actions",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "node|user|token|ch_table|...",
                        "name": "target_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "фильтр по target_id",
                        "name": "target_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 (начало)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 (конец)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 100, max 1000",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ListAuditResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/audit/export.csv": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "ApiTokenAuth": []
                    }
                ],
                "description": "Тот же набор фильтров, что у /api/audit. Default limit 10000, max 50000. CSV в UTF-8 с BOM (для Excel), 9 колонок: id, created_at, user_login, user_id, action, target_type, target_id, ip_address, details (JSON).",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "Выгрузка audit-log в CSV (admin only, §7.13).",
                "parameters": [
                    {
                        "type": "string",
                        "description": "фильтр по user_id",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "одно значение action",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "несколько action через запятую",
                        "name": "actions",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "node|user|token|ch_table|...",
                        "name": "target_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "фильтр по target_id",
                        "name": "target_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 (начало)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 (конец)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 10000, max 50000",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_web_adapter_in_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "password_change_required"
                },
                "error": {
                    "type": "string",
                    "example": "not found"
                }
            }
        },
...
        "internal_web_adapter_in_http.ListAuditResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_web_adapter_in_http.auditEntryResponse"
                    }
                }
            }
        },
...
        "internal_web_adapter_in_http.auditEntryResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "id": {
                    "type": "string"
                },
                "ip_address": {
                    "type": "string"
                },
                "target_id": {
                    "type": "string"
                },
                "target_type": {
                    "type": "string"
                },
                "team_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "user_login": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiTokenAuth": {
            "description": "\"Bearer db_\u003c32-байт-base64\u003e\". См. §7.14.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "nexus_session",
            "in": "cookie"
        }
    }
}